Resolved #2366 by having the ascii draw methods be method extensions of Tile

Updated the CHANGELOG


Fixed bug
